player controller

+++ SoccerProject/src/main/java/com/itacademy/soccer/dao/IPlayerDAO.java
@@ -12,7 +12,7 @@ public interface IPlayerDAO extends JpaRepository<Player, Long> {
 	
 	Player findByName(String name);
 
-	List<Player> findByteamId(int teamId);
+	List<Player> findByteamId(Long teamId);
 	
 	
 }
